refactor(examples): separate presentation and shaders
