+++ examples/api/cubemap/app-ui.ts
@@ -1,0 +1,6 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const CUBEMAP_INFO_HTML =
+  'Uses a luma.gl <code>TextureCube</code> to simulate a reflective surface';
