+++ examples/api/render-bundles/app.ts
@@ -7,72 +7,23 @@ import {Buffer, Device, normalizeBindingsByGroup} from '@luma.gl/core';
 import type {AnimationProps} from '@luma.gl/engine';
 import {AnimationLoopTemplate, CubeGeometry, Model, OrbitControls} from '@luma.gl/engine';
 import {Matrix4, radians} from '@math.gl/core';
-import {
-  ColumnPanel,
-  type Panel,
-  type SettingsChangeDescriptor,
-  type SettingsSchema
-} from '@deck.gl-community/panels';
+import {type SettingsChangeDescriptor} from '@deck.gl-community/panels';
 import {
   ExamplePanelManager,
   ExampleSettingsPanelManager,
   getChangedSetting,
-  makeExamplePanelHostHtml,
-  makeHtmlCustomPanel
+  makeExamplePanelHostHtml
 } from '../../example-panels';
+import {
+  DEFAULT_DRAW_COUNT,
+  isRenderBundleDrawCount,
+  makeRenderBundlesSettingsSchema,
+  RenderBundlesUI
+} from './app-ui';
+import {CAMERA_DISTANCE, CAMERA_PITCH, CAMERA_YAW, makeObjectUniforms} from './scene';
+import {WGSL_SHADER} from './shaders';
 
-const DEFAULT_DRAW_COUNT = 5000;
-const DRAW_COUNT_OPTIONS = [1000, DEFAULT_DRAW_COUNT, 10000];
-const DRAW_COUNT_OPTION_SET = new Set(DRAW_COUNT_OPTIONS);
 const FRAME_UNIFORM_FLOAT_COUNT = 32;
-const OBJECT_UNIFORM_FLOAT_COUNT = 20;
-const CAMERA_DISTANCE = Math.hypot(38, 12);
-const CAMERA_PITCH = Math.atan2(12, 38);
-const CAMERA_YAW = Math.PI / 2;
-
-const WGSL_SHADER = /* wgsl */ `\
-struct FrameUniforms {
-  viewMatrix: mat4x4<f32>,
-  projectionMatrix: mat4x4<f32>,
-};
-
-@group(0) @binding(0) var<uniform> frameUniforms : FrameUniforms;
-
-struct ObjectUniforms {
-  modelMatrix: mat4x4<f32>,
-  color: vec4<f32>,
-};
-
-@group(1) @binding(0) var<uniform> objectUniforms : ObjectUniforms;
-
-struct VertexInputs {
-  @location(0) positions : vec4<f32>,
-  @location(1) normals : vec3<f32>,
-};
-
-struct FragmentInputs {
-  @builtin(position) position : vec4<f32>,
-  @location(0) normal : vec3<f32>,
-  @location(1) color : vec4<f32>,
-};
-
-@vertex
-fn vertexMain(inputs: VertexInputs) -> FragmentInputs {
-  var outputs: FragmentInputs;
-  let worldPosition = objectUniforms.modelMatrix * inputs.positions;
-  outputs.position = frameUniforms.projectionMatrix * frameUniforms.viewMatrix * worldPosition;
-  outputs.normal = normalize((objectUniforms.modelMatrix * vec4<f32>(inputs.normals, 0.0)).xyz);
-  outputs.color = objectUniforms.color;
-  return outputs;
-}
-
-@fragment
-fn fragmentMain(inputs: FragmentInputs) -> @location(0) vec4<f32> {
-  let lightDirection = normalize(vec3<f32>(0.4, 0.8, 0.3));
-  let diffuse = max(dot(normalize(inputs.normal), lightDirection), 0.18);
-  return vec4<f32>(inputs.color.rgb * diffuse, inputs.color.a);
-}
-`;
 
 type SceneRenderable = {
   bindings: Bindings;
@@ -91,14 +42,12 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
   readonly frameUniformBuffer: Buffer;
   readonly settingsPanel: ExampleSettingsPanelManager;
   readonly panels: ExamplePanelManager;
+  readonly ui: RenderBundlesUI;
   private readonly frameBindGroupCacheKey = {};
   renderables: SceneRenderable[] = [];
   renderBundle: RenderBundle | null = null;
   drawCount = DEFAULT_DRAW_COUNT;
   useRenderBundles = true;
-  private cpuTimeElement: HTMLElement | null = null;
-  private drawCountElement: HTMLElement | null = null;
-  private modeElement: HTMLElement | null = null;
   private orbitControls: OrbitControls | null = null;
 
   constructor({device}: AnimationProps) {
@@ -134,7 +83,8 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
       },
       onSettingsChange: this.handleSettingsChange
     });
-    this.panels = new ExamplePanelManager({panel: this.makePanel()});
+    this.ui = new RenderBundlesUI(this.settingsPanel, this.drawCount);
+    this.panels = new ExamplePanelManager({panel: this.ui.panel});
     this.createScene();
     this.rebuildRenderBundle();
     this.panels.mount();
@@ -171,7 +121,11 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
       this.renderScene(renderPass);
     }
     renderPass.end();
-    this.updateStats(animationLoop.cpuTime.getSampleAverageTime());
+    this.ui.updateStats(
+      animationLoop.cpuTime.getSampleAverageTime(),
+      this.drawCount,
+      this.useRenderBundles
+    );
   }
 
   onFinalize(): void {
@@ -270,72 +224,6 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
     this.frameUniformBuffer.write(frameUniforms);
   }
 
-  private makePanel(): Panel {
-    return new ColumnPanel({
-      id: 'api-render-bundles-controls',
-      title: 'Render Bundles',
-      panels: [
-        makeHtmlCustomPanel({
-          id: 'api-render-bundles-description',
-          title: '',
-          html: `\
-          <p>This scene intentionally issues thousands of individual draws. Render bundles record those WebGPU commands once, so per-frame CPU work is mostly the camera buffer update and bundle replay.</p>
-          `
-        }),
-        makeHtmlCustomPanel({
-          id: 'api-render-bundles-stats',
-          title: 'Frame Stats',
-          html: `\
-          <style>
-            .render-bundle-stats {
-              display: grid;
-              gap: 8px;
-            }
-            .render-bundle-stat {
-              display: flex;
-              justify-content: space-between;
-              gap: 16px;
-              font-variant-numeric: tabular-nums;
-            }
-            .render-bundle-stat strong {
-              font-weight: 700;
-            }
-          </style>
-          <div class="render-bundle-stats">
-            <div class="render-bundle-stat"><span>CPU Time</span><strong data-cpu-time>0.00 ms</strong></div>
-            <div class="render-bundle-stat"><span>Draw Calls</span><strong data-draw-count>${formatDrawCount(this.drawCount)}</strong></div>
-            <div class="render-bundle-stat"><span>Mode</span><strong data-mode>Render bundle</strong></div>
-          </div>
-          `,
-          onRender: rootElement => {
-            this.cpuTimeElement = rootElement.querySelector('[data-cpu-time]');
-            this.drawCountElement = rootElement.querySelector('[data-draw-count]');
-            this.modeElement = rootElement.querySelector('[data-mode]');
-            this.updateStats(0);
-            return () => {
-              this.cpuTimeElement = null;
-              this.drawCountElement = null;
-              this.modeElement = null;
-            };
-          }
-        }),
-        this.settingsPanel.makePanel()
-      ]
-    });
-  }
-
-  private updateStats(cpuTimeMilliseconds: number): void {
-    if (this.cpuTimeElement) {
-      this.cpuTimeElement.textContent = `${cpuTimeMilliseconds.toFixed(2)} ms`;
-    }
-    if (this.drawCountElement) {
-      this.drawCountElement.textContent = formatDrawCount(this.drawCount);
-    }
-    if (this.modeElement) {
-      this.modeElement.textContent = this.useRenderBundles ? 'Render bundle' : 'Replay draws';
-    }
-  }
-
   private readonly handleSettingsChange = (
     _settings: Record<string, unknown>,
     changedSettings?: SettingsChangeDescriptor[]
@@ -343,7 +231,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
     const useRenderBundles = getChangedSetting(changedSettings, 'useRenderBundles')?.nextValue;
     if (typeof useRenderBundles === 'boolean') {
       this.useRenderBundles = useRenderBundles;
-      this.updateStats(0);
+      this.ui.updateStats(0, this.drawCount, this.useRenderBundles);
     }
 
     const drawCount = getChangedSetting(changedSettings, 'drawCount')?.nextValue;
@@ -353,7 +241,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
   };
 
   private handleDrawCountChange(drawCount: number): void {
-    if (!DRAW_COUNT_OPTION_SET.has(drawCount) || drawCount === this.drawCount) {
+    if (!isRenderBundleDrawCount(drawCount) || drawCount === this.drawCount) {
       return;
     }
 
@@ -361,80 +249,6 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
     this.destroyScene();
     this.createScene();
     this.rebuildRenderBundle();
-    this.updateStats(0);
+    this.ui.updateStats(0, this.drawCount, this.useRenderBundles);
   }
-}
-
-function makeRenderBundlesSettingsSchema(): SettingsSchema {
-  return {
-    title: 'Settings',
-    sections: [
-      {
-        id: 'render-bundles',
-        name: 'Render',
-        initiallyCollapsed: false,
-        settings: [
-          {
-            name: 'useRenderBundles',
-            label: 'Use Render Bundles',
-            type: 'boolean',
-            persist: 'none'
-          },
-          {
-            name: 'drawCount',
-            label: 'Draw Calls',
-            type: 'select',
-            persist: 'none',
-            options: DRAW_COUNT_OPTIONS.map(drawCount => ({
-              label: formatDrawCount(drawCount),
-              value: drawCount
-            }))
-          }
-        ]
-      }
-    ]
-  };
-}
-
-function makeObjectUniforms(index: number, drawCount: number): Float32Array {
-  const objectUniforms = new Float32Array(OBJECT_UNIFORM_FLOAT_COUNT);
-  const modelMatrix = makeModelMatrix(index, drawCount);
-  objectUniforms.set(modelMatrix.toArray(), 0);
-  objectUniforms.set(makeObjectColor(index), 16);
-  return objectUniforms;
-}
-
-function makeModelMatrix(index: number, drawCount: number): Matrix4 {
-  if (index === 0) {
-    return new Matrix4().scale([5, 5, 5]);
-  }
-
-  const normalizedIndex = index / Math.max(drawCount - 1, 1);
-  const angle = normalizedIndex * Math.PI * 96 + index * 0.013;
-  const radius = 10 + (index % 64) * 0.14;
-  const height = Math.sin(index * 0.73) * 2.4;
-  const scale = 0.08 + ((index * 17) % 23) / 110;
-
-  return new Matrix4()
-    .translate([Math.cos(angle) * radius, height, Math.sin(angle) * radius])
-    .rotateXYZ([index * 0.09, index * 0.13, index * 0.17])
-    .scale([scale, scale, scale]);
-}
-
-function makeObjectColor(index: number): [number, number, number, number] {
-  if (index === 0) {
-    return [0.86, 0.76, 0.5, 1];
-  }
-
-  const colorBand = index % 7;
-  return [
-    0.28 + colorBand * 0.055,
-    0.24 + ((index * 3) % 7) * 0.045,
-    0.3 + ((index * 5) % 7) * 0.05,
-    1
-  ];
-}
-
-function formatDrawCount(drawCount: number): string {
-  return `${drawCount.toLocaleString()} individual draws`;
 }
