+++ examples/api/texture-3d/app-ui.ts
@@ -1,0 +1,8 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const TEXTURE_3D_INFO_HTML = `
+<p>Volumetric 3D noise visualized using a <b>3D texture</b>.</p>
+<p>Uses the luma.gl <code>Texture3D</code> class.</p>
+`;
