+++ examples/api/texture-compressed/app-ui.ts
@@ -1,0 +1,11 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const TEXTURE_COMPRESSED_INFO_HTML = `
+<p>Drawing a compressed texture</p>
+<p>
+  Rendered using the luma.gl <code>Model</code>, <code>CubeGeometry</code> and
+  <code>AnimationLoop</code> classes.
+</p>
+`;
