+++ examples/experimental/a-buffer/app.ts
@@ -61,144 +61,7 @@ const scene = {
   }
 } as const satisfies ShaderModule<SceneUniforms>;
 
-const SCENE_WGSL = /* wgsl */ `\
-struct SceneUniforms {
-  viewProjectionMatrix: mat4x4<f32>,
-  opacity: f32,
-  sphereSize: f32,
-  gridStart: u32,
-  gridStride: u32,
-};
-
-@group(0) @binding(auto) var<uniform> scene: SceneUniforms;
-
-struct VertexInputs {
-  @builtin(instance_index) instanceIndex: u32,
-  @location(0) positions: vec3<f32>,
-  @location(1) normals: vec3<f32>,
-};
-
-struct FragmentInputs {
-  @builtin(position) Position: vec4<f32>,
-  @location(0) normal: vec3<f32>,
-  @location(1) @interpolate(flat) gridIndex: u32,
-};
-
-@vertex
-fn vertexMain(inputs: VertexInputs) -> FragmentInputs {
-  let gridIndex = scene.gridStart + inputs.instanceIndex * scene.gridStride;
-  let row = gridIndex / 3u;
-  let column = gridIndex % 3u;
-  let offset = vec3<f32>(
-    (f32(column) - 1.0) * 9.0,
-    (1.0 - f32(row)) * 8.0,
-    0.0
-  );
-
-  var outputs: FragmentInputs;
-  outputs.Position = scene.viewProjectionMatrix * vec4<f32>(inputs.positions * scene.sphereSize + offset, 1.0);
-  outputs.normal = inputs.normals;
-  outputs.gridIndex = gridIndex;
-  return outputs;
-}
-
-@fragment
-fn fragmentMain(inputs: FragmentInputs) -> @location(0) vec4<f32> {
-  let colors = array<vec3<f32>, 6>(
-    vec3<f32>(0.96, 0.18, 0.24),
-    vec3<f32>(0.20, 0.78, 0.36),
-    vec3<f32>(0.16, 0.48, 0.98),
-    vec3<f32>(0.86, 0.22, 0.92),
-    vec3<f32>(1.0, 0.72, 0.12),
-    vec3<f32>(0.08, 0.78, 0.88)
-  );
-  let baseColor = colors[inputs.gridIndex % 6u];
-  let lightDirection = normalize(vec3<f32>(0.7, 1.0, -0.4));
-  let lighting = 0.36 + 0.64 * (0.5 + 0.5 * dot(normalize(inputs.normal), lightDirection));
-  let color = vec4<f32>(baseColor * lighting, scene.opacity);
-
-#if A_BUFFER_ENABLED
-  return aBuffer_captureStraightColor(color, inputs.Position);
-#else
-#if WBOIT_ENABLED
-  return wboit_captureStraightColor(color, inputs.Position);
-#else
-  return color;
-#endif
-#endif
-}
-`;
-
-const SCENE_GLSL = /* glsl */ `\
-#version 300 es
-precision highp float;
-precision highp int;
-
-uniform sceneUniforms {
-  mat4 viewProjectionMatrix;
-  float opacity;
-  float sphereSize;
-  uint gridStart;
-  uint gridStride;
-} scene;
-
-in vec3 positions;
-in vec3 normals;
-
-out vec3 vNormal;
-flat out uint vGridIndex;
-
-void main(void) {
-  uint gridIndex = scene.gridStart + uint(gl_InstanceID) * scene.gridStride;
-  uint row = gridIndex / 3u;
-  uint column = gridIndex % 3u;
-  vec3 offset = vec3((float(column) - 1.0) * 9.0, (1.0 - float(row)) * 8.0, 0.0);
-
-  gl_Position = scene.viewProjectionMatrix * vec4(positions * scene.sphereSize + offset, 1.0);
-  vNormal = normals;
-  vGridIndex = gridIndex;
-}
-`;
-
-const SCENE_FS_GLSL = /* glsl */ `\
-#version 300 es
-precision highp float;
-precision highp int;
-
-uniform sceneUniforms {
-  mat4 viewProjectionMatrix;
-  float opacity;
-  float sphereSize;
-  uint gridStart;
-  uint gridStride;
-} scene;
-
-in vec3 vNormal;
-flat in uint vGridIndex;
-
-out vec4 fragColor;
-
-void main(void) {
-  vec3 colors[6] = vec3[6](
-    vec3(0.96, 0.18, 0.24),
-    vec3(0.20, 0.78, 0.36),
-    vec3(0.16, 0.48, 0.98),
-    vec3(0.86, 0.22, 0.92),
-    vec3(1.0, 0.72, 0.12),
-    vec3(0.08, 0.78, 0.88)
-  );
-  vec3 baseColor = colors[vGridIndex % 6u];
-  vec3 lightDirection = normalize(vec3(0.7, 1.0, -0.4));
-  float lighting = 0.36 + 0.64 * (0.5 + 0.5 * dot(normalize(vNormal), lightDirection));
-  vec4 color = vec4(baseColor * lighting, scene.opacity);
-
-#if WBOIT_ENABLED
-  fragColor = wboit_captureStraightColor(color, gl_FragCoord);
-#else
-  fragColor = color;
-#endif
-}
-`;
+import {SCENE_FS_GLSL, SCENE_GLSL, SCENE_WGSL} from './shaders';
 
 const OPAQUE_PARAMETERS = {
   depthWriteEnabled: true,
@@ -599,20 +462,7 @@ function isTransparencyMode(value: unknown): value is TransparencyMode {
   return value === 'a-buffer' || value === 'weighted-blended' || value === 'alpha-blending';
 }
 
-const A_BUFFER_DESCRIPTION_HTML = `\
-<p>Compare exact per-pixel linked-list OIT, approximate weighted blended OIT, and unsorted alpha blending on the same interleaved scene.</p>
-<p>The dropdown renders the same unsorted translucent spheres with three different blending strategies.</p>
-<div style="display: grid; gap: 8px;">
-  <div><strong>A-buffer OIT</strong><br>Captures every translucent fragment into GPU storage, sorts fragments per pixel, then composites them back-to-front. It is the most accurate option here and is available on WebGPU.</div>
-  <div><strong>Weighted blended OIT</strong><br>Accumulates weighted color and revealage into floating-point offscreen targets, then composites once. It avoids sorting and works on WebGPU or WebGL2 when float render-target blending is available, but it is approximate.</div>
-  <div><strong>Standard alpha blending</strong><br>Blends translucent draws directly into the framebuffer in submission order. It is the cheapest fallback, but unsorted overlaps produce visibly incorrect results.</div>
-</div>`;
-
-const A_BUFFER_BACKGROUND_HTML = `\
-<p><b>Why transparency is hard:</b> ordinary alpha blending is order dependent. Correct color requires fragments for the same pixel to be composited back-to-front, but draw submission order rarely matches depth order for intersecting translucent geometry.</p>
-<p><b>A-buffer OIT:</b> the capture pass stores each translucent fragment's color and depth in GPU storage. A resolve pass gathers the fragments for one pixel, sorts them by depth, and composites them back-to-front. The result is exact up to the configured per-pixel storage budget.</p>
-<p><b>Weighted blended OIT:</b> instead of sorting, two blended render targets accumulate weighted color and revealage. The final resolve estimates the transparent stack in one pass. It is usually faster and more portable, but dense overlaps can reveal the approximation.</p>
-<p><b>Performance trade-off:</b> A-buffer quality costs storage, atomics, and per-pixel sorting; weighted blending costs extra attachments and bandwidth; ordinary alpha blending is cheapest but makes correctness the caller's draw-order problem.</p>`;
+import {A_BUFFER_BACKGROUND_HTML, A_BUFFER_DESCRIPTION_HTML} from './app-ui';
 
 function createSceneFramebuffer(device: Device, width: number, height: number): Framebuffer {
   const colorTexture = device.createTexture({
