+++ examples/showcase/algebraic-varieties/app-ui.ts
@@ -1,0 +1,10 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const ALGEBRAIC_VARIETIES_INFO_HTML = `
+<section class="variety-info">
+  <p>Real algebraic surfaces are evaluated and intersected directly in WGSL—there is no mesh and the polynomial is not treated as an SDF.</p>
+  <p>Drag to orbit · wheel to zoom · <strong>1–9, 0</strong> presets · <strong>S</strong> singularities · <strong>R</strong> reset</p>
+  <div class="variety-badges"><span>WebGPU</span><span>analytic gradients</span><span>HDR lighting</span><span>hybrid root refinement</span></div>
+</section>`;
