+++ examples/showcase/dof/app-ui.ts
@@ -1,0 +1,10 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const DOF_DESCRIPTION_HTML = `\
+<div>
+  <p><b>Depth of Field</b></p>
+  <p>Several instanced cubes rendered with a depth-driven postprocessing blur.</p>
+</div>
+`;
