+++ examples/showcase/globe/app-ui.ts
@@ -1,0 +1,9 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const GLOBE_DESCRIPTION_HTML = `\
+<p>Revives the classic Earth specular demo as a modern luma.gl v9 showcase with animated oceans and a starfield backdrop.</p>
+<p>Drag to orbit. Use the mouse wheel or trackpad to zoom.</p>
+<p style="font-size: 11px;">Sky: NASA <a href="https://science.nasa.gov/3d-resources/tycho-star-map/">Tycho Star Map</a>.</p>
+`;
