+++ examples/showcase/persistence/app-ui.ts
@@ -1,0 +1,10 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const PERSISTENCE_INFO_HTML = `
+<p>Electron trails renderings persist across multiple frames.</p>
+<p>
+  Uses multiple luma.gl <code>Framebuffer</code>s to hold previously rendered data between frames.
+</p>
+`;
