+++ examples/tutorials/hello-triangle/app-ui.ts
@@ -1,0 +1,5 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const HELLO_TRIANGLE_INFO_HTML = '<p>Have to start somewhere...</p>';
