+++ examples/tutorials/lighting/app-ui.ts
@@ -1,0 +1,7 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const LIGHTING_INFO_HTML = `
+<p>Drawing a phong-shaded cube with multiple colored spot lights aimed at the cube</p>
+`;
