+++ examples/tutorials/transform/app-ui.ts
@@ -1,0 +1,14 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: Copyright (c) vis.gl contributors
+
+export const TRANSFORM_INFO_HTML = `
+<p>
+  Instanced triangles animated on the GPU using a luma.gl <code>BufferTransform</code> object.
+
+  This is a port of an example from
+  <a href="https://github.com/WebGLSamples/WebGL2Samples/blob/master/samples/transform_feedback_instanced.html">
+    WebGL2Samples
+  </a>
+</p>
+`;
